requirements.txt
aioredis==1.3.1
amqp==2.5.2
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.1.3
Automat==0.8.0
billiard==3.6.2.0
celery==4.4.0
certifi==2019.11.28
cffi==1.13.2
channels==2.4.0
channels-redis==2.4.1
chardet==3.0.4
constantly==15.1.0
cryptography==2.8
daphne==2.4.1
Django==3.0.3
django-filter==2.2.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
hiredis==1.0.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
kombu==4.6.7
msgpack==0.6.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
PyHamcrest==2.0.0
PyJWT==1.7.1
pyOpenSSL==19.1.0
pytz==2019.3
redis==3.4.1
requests==2.22.0
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.0
Twisted==19.10.0
txaio==20.1.1
urllib3==1.25.8
vine==1.3.0
zope.interface==4.7.1
